// src/app/model/gestionClientes/listadoclientes.ts
export class ListadoClientes {
  id: number = 0;
  nombres: string = '';
  apellidos: string = '';
  tipodocident: number = 0;
  nrodocident: string = '';
  email: string = '';
  telefono: string = '';
  codcalle: number = 0;
  nrocalle: string = '';
  referencia: string = '';
  latitud: string = '';
  longitud: string = '';
  estareg: number = 0;
  fechareg: string = '';
  creador: string = '';
  fechabaja: string = '';
  observacion_baja: string | null = null;
  destipocalle: string = '';
  descripcioncalle: string = '';
  descripcioncorta: string = '';
  destipodocident: string = '';
  tipodocidenabr: string = '';
}