// src/app/model/gestionCobranza/ListadomorosidadRequest.ts
export class ListadomorosidadRequest {
  nombre_completo: string = '';
  fecharebaja: string = '';
  id_baja: number = 0;
  id_cliente: number = 0;
  fechareg: string = '';
  estado: string = '';
  observacion: string = '';
  motivo: string = '';
  desctipo: string = '';
  email: string = '';
  nrodocident: string = '';
  telefono: string = '';
}